requirements.txt
# ---------- accel --------- #
"jax[cuda12_pip]==0.4.14" --find-links "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
pip3 install torch --index-url https://download.pytorch.org/whl/cpu  # to use DataLoader
equinox
optax
evosax
evojax
tensorboard  # to log experiments

# --------- hydra --------- #
hydra-core==1.3.1
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
hydra-zen>=0.9.0

#---------- others -----------#
requests
scipy>=1.10.0
matplotlib>=3.5.2
scikit-learn>=1.0.2
scikit-image>=0.19.2
seaborn>=0.11.2
jupyter>=1.0.0
h5py>=3.6.0
pyrootutils     # standardizing the project root setup
# pre-commit    # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
